Rename settings.lua to mcp_settings.lua, workaround for https://github.com/Blightmud/Blightmud/issues/1059
