+++ src/mcp_settings.lua
